requirements.txt
dash==2.18.2
plotly==5.24.1
pandas==2.2.3
numpy==2.1.3
python-dateutil==2.9.0
